refactor(overlay): dispatch the reconcile poller per workspace
The poller walked every overlay-mode workspace's due connection and swept each
inline, logging and skipping a failure — so a workspace whose incumbent was
rate-limited, revoked or unreachable left no durable trace while River recorded
the tick as completed. Each workspace now succeeds or fails as its own row, and
the sweep backoff it records is unchanged.

The fan-out lands on a SERIAL queue. overlaybudget.ConsumeSearch counts but
does not pace, and its keys are per workspace, so it cannot bound a
provider-level burst: five connections reconciling concurrently could exceed
the incumbent's per-second Search limit. Per-workspace visibility is the goal
here; per-workspace parallelism is not.

The job names only the workspace. Its incumbent, region, credential ref and
connected-at are re-read at work time through overlay.ActiveConnection — the
per-request counterpart to the fleet walk — so a connection that changed
between the due-scan and the sweep is swept under its CURRENT identity, and
the disconnect+reconnect fence narrows rather than widens.

Two suites changed their assertions, and only because of the one behavioural
change this phase is for: a connection-level failure now fails the workspace's
job row instead of being swallowed. The mid-sweep-disconnect clean stop still
returns nil, because an aborted fenced write is not a failure to retry.

jobs_overlay.go crossed the 500-line cap, so the per-object-class sweep phases
move to jobs_overlay_sweep.go.
